Feat/user login

+++ repet-back/repet/views.py
@@ -4,21 +4,70 @@ from django.http.response import Http404
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.views import APIView
-from .serializers import UserSerializer, PetSerializer, PetSerializerGET, VaccineSerializer, VaccineSerializerGET, RecordSerializer, RecordSerializerGET, RemindersSerializer, RemindersSerializerGET
+from .serializers import UserSerializer, PetSerializer, PetSerializerGET, VaccineSerializer, VaccineSerializerGET, RecordSerializer, RecordSerializerGET, RemindersSerializer, RemindersSerializerGET, RegisterSerializer
 from .models import User, Pet, Vaccine, Record, Reminder
+from django.contrib.auth.models import User as USER
+from rest_framework.authentication import SessionAuthentication, BasicAuthentication
+from rest_framework.permissions import IsAuthenticated
 
 # Create your views here.
 
-class UserView(APIView):
+from rest_framework.authtoken.views import ObtainAuthToken
+from rest_framework.authtoken.models import Token
+from rest_framework.response import Response
+import datetime
+from django.utils.timezone import utc
+
+from .authentication import authenticate_credentials
+
+class CustomAuthToken(ObtainAuthToken):
+    def post(self, request, *args, **kwargs):
+        serializer = self.serializer_class(data=request.data,
+                                           context={'request': request})
+        
+        if '@' in request.data['username']:
+            request.data['username'] = User.objects.get(email=request.data['username']).username
+            
+        serializer.is_valid(raise_exception=True)
+        user = serializer.validated_data['user']
+        token, created = Token.objects.get_or_create(user=user)
+
+        return Response({
+            'token': token.key,
+            'user_id': User.objects.get(user_login=user.pk).pk,
+            'email': user.email
+        })
+    
+
+
+def get_my_id(user_pk):
+    return User.objects.get(user_login=user_pk)
+
+def can_acess(user_pk, pk):
+    return get_my_id(user_pk).pk == pk
+
+class CheckTokenView(APIView):
+    def get(self, request):
+        try:
+            return Response({"detail": authenticate_credentials(request.auth)})
+        except:
+            return Response({"detail": False})
+
+class RegisterView(APIView):
+    permission_classes = []
     # create
     def post(self, request):
-        serializer = UserSerializer(data=request.data)
+        serializer = RegisterSerializer(data=request.data)
 
         if serializer.is_valid():
             serializer.save()
             return Response("Usuário cadastrado com sucesso.", status=status.HTTP_201_CREATED)
         else:
+            # print(serializer.error_messages)
             return Response("Erro ao cadastrar o usuário.", status=status.HTTP_400_BAD_REQUEST)
+
+class UserView(APIView):
+    permission_classes = [IsAuthenticated]
         
     def query_user(self, pk):
         try:
@@ -30,6 +79,8 @@ class UserView(APIView):
     # read
     def get(self, request, pk=None):
         if pk:
+            if not can_acess(request.user.id, pk):
+                return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
             data = self.query_user(pk)
             serializer = UserSerializer(data)
         else:
@@ -39,23 +90,46 @@ class UserView(APIView):
     
     # update
     def put(self, request, pk):
+        if not can_acess(request.user.id, pk):
+            return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
         user_update = self.query_user(pk)
         serializer = UserSerializer(instance=user_update, data=request.data, partial=True)
 
+        #print(serializer.initial_data['email'])
         if serializer.is_valid():
+            if serializer.initial_data.get('email') or serializer.initial_data.get('username'):
+                us1 = USER.objects.filter(username = serializer.initial_data.get('username'))
+                us2 = USER.objects.filter(email = serializer.initial_data.get('email'))
+
+                if us1.exists() or us2.exists():
+                    Response("Erro ao atualizar o usuário.", status=status.HTTP_400_BAD_REQUEST)
             serializer.save()
+            us = USER.objects.filter(pk=user_update.user_login.pk)
+
+            #print(us)
+            if serializer.initial_data.get('email'):
+                us.update(email=serializer.initial_data.get('email'))
+            
+            if serializer.initial_data.get('username'):
+                us.update(username=serializer.initial_data.get('username'))
             return Response("Usuário atualizado com sucesso.", status=status.HTTP_204_NO_CONTENT)
         else:
             return Response("Erro ao atualizar o usuário.", status=status.HTTP_400_BAD_REQUEST)
     
     # delete
     def delete(self, request, pk):
+        if not can_acess(request.user.id, pk):
+            return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
         user_delete = self.query_user(pk)
-        user_delete.delete()
+
+        user = user_delete.user_login
+        user.delete()
+        # user_delete.delete()
 
         return Response("Usuário deletado com sucesso.")
 
 class PetView(APIView):
+    permission_classes = [IsAuthenticated]
 
     def post(self, request):
         serializer = PetSerializer(data=request.data)
@@ -72,11 +146,16 @@ class PetView(APIView):
             raise Http404
 
     def get(self, request, pk=None):
+        id = get_my_id(request.user.id)
         if pk:
             data = self.query_pet(pk)
+            if not can_acess(request.user.id, data.user.pk):
+                return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
+            # data = data.filter(user=id)
             serializer = PetSerializerGET(data)
         else:
             data = Pet.objects.all()
+            data = data.filter(user=id)
             serializer = PetSerializerGET(data, many=True)
 
         response = Response(serializer.data)
@@ -85,6 +164,8 @@ class PetView(APIView):
     
     def put(self, request, pk):
         pet_update = self.query_pet(pk)
+        if not can_acess(request.user.id, pet_update.user.pk):
+            return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
         serializer = PetSerializer(instance=pet_update, data=request.data, partial=True)
 
         if serializer.is_valid():
@@ -95,11 +176,14 @@ class PetView(APIView):
             
     def delete(self, request, pk):
         pet_delete = self.query_pet(pk)
+        if not can_acess(request.user.id, pet_delete.user.pk):
+            return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
         pet_delete.delete()
 
         return Response("Pet deletado com sucesso.")  
 
 class RecordView(APIView):
+    permission_classes = [IsAuthenticated]
     def post(self, request):
         serializer = RecordSerializer(data=request.data)
 
@@ -118,13 +202,20 @@ class RecordView(APIView):
     def get(self, request, pk=None):
         # list just the records, without vaccines
         # get id records of all vaccines
+        id = get_my_id(request.user.id)
+        
         list_ids = [x.record.id for x in Vaccine.objects.all()]
+        list_pets = [x.id for x in Pet.objects.filter(user=id)]
 
         if pk:
             data = self.query_record(pk)
+            if not can_acess(request.user.id, data.pet.user.pk):
+                return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
+            # data = data.filter(user=id)
             serializer = RecordSerializerGET(data)
         else:
             data = Record.objects.all()
+            data = data.filter(pet__in=list_pets)
             if request.GET.get("pet_id"):
                 data = data.filter(pet=request.GET.get("pet_id"))
             data = data.exclude(id__in=list_ids)
@@ -133,6 +224,8 @@ class RecordView(APIView):
     
     def put(self, request, pk):
         record_update = self.query_record(pk)
+        if not can_acess(request.user.id, record_update.pet.user.pk):
+            return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
         serializer = RecordSerializer(instance=record_update, data=request.data, partial=True)
 
         if serializer.is_valid():
@@ -143,19 +236,14 @@ class RecordView(APIView):
             
     def delete(self, request, pk):
         record_delete = self.query_record(pk)
+        if not can_acess(request.user.id, record_delete.pet.user.pk):
+            return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
         record_delete.delete()
 
         return Response("Registro deletado com sucesso.")
 
-class MyPetView(APIView):
-    
-    def get(self, request, pk):
-        pets = Pet.objects.filter(user_id=pk)
-        serializer = PetSerializer(pets, many=True)
-
-        return Response(serializer.data)
-
 class VaccineView(APIView):
+    permission_classes = [IsAuthenticated]
 
     def post(self, request):
         records_serializes = RecordSerializer(data=request.data)
@@ -187,18 +275,27 @@ class VaccineView(APIView):
             return Http404
 
     def get(self, request, pk=None):
+        id = get_my_id(request.user.id)
+        list_pets = [x.id for x in Pet.objects.filter(user=id)]
         if pk:
             data = self.query_vaccine(pk)
+            if not can_acess(request.user.id, data.pet.user.pk):
+                return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
+            # data = data.filter(pet__in=list_pets)
             serializer = VaccineSerializerGET(data)
         else:
             data = Vaccine.objects.all()
+            data = data.filter(pet__in=list_pets)
             if request.GET.get("pet_id"):
                 data = data.filter(pet=request.GET.get("pet_id"))
             serializer = VaccineSerializerGET(data, many=True)
         return Response(serializer.data)
 
     def put(self, request, pk):
+
         vacc_query = self.query_vaccine(pk)
+        if not can_acess(request.user.id, vacc_query.pet.user.pk):
+            return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
         record_update = self.query_records(vacc_query.record.id)
         serializer = RecordSerializer(instance=record_update, data=request.data, partial=True)
 
@@ -218,12 +315,15 @@ class VaccineView(APIView):
     def delete(self, request, pk=None):
         if pk:
             vaccine_delete = self.query_vaccine(pk)
+            if not can_acess(request.user.id, vaccine_delete.pet.user.pk):
+                return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
             vaccine_delete.delete()
             return Response("Vacina deletada com sucesso.")
         else: 
             return Response("Erro ao deletar vacinas.", status=status.HTTP_400_BAD_REQUEST)
 
 class RemindersView(APIView):
+    permission_classes = [IsAuthenticated]
 
     def post(self, request):
         serializer = RemindersSerializer(data=request.data)
@@ -241,11 +341,18 @@ class RemindersView(APIView):
             return Http404
 
     def get(self, request, pk=None):
+        id = get_my_id(request.user.id)
+
+        list_pets = [x.id for x in Pet.objects.filter(user=id)]
         if pk:
             data = self.query_reminders(pk)
+            if not can_acess(request.user.id, data.pet.user.pk):
+                return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
+            # data = data.filter(pet__in=list_pets)
             serializer = RemindersSerializerGET(data)
         else:
             data = Reminder.objects.all()
+            data = data.filter(pet__in=list_pets)
             if request.GET.get("pet_id"):
                 data = data.filter(pet=request.GET.get("pet_id"))
             serializer = RemindersSerializerGET(data, many=True)
@@ -253,6 +360,8 @@ class RemindersView(APIView):
 
     def put(self, request, pk):
         reminders_update = self.query_reminders(pk)
+        if not can_acess(request.user.id, reminders_update.pet.user.pk):
+            return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
         serializer = RemindersSerializer(instance=reminders_update, data=request.data, partial=True)
 
         if serializer.is_valid():
@@ -263,6 +372,8 @@ class RemindersView(APIView):
             
     def delete(self, request, pk):
         reminders_delete = self.query_reminders(pk)
+        if not can_acess(request.user.id, reminders_delete.pet.user.pk):
+            return Response("Não autorizado", status=status.HTTP_401_UNAUTHORIZED)
         reminders_delete.delete()
 
         return Response("Lembrete deletado com sucesso.")
